Added Annotation @RequestMapping.

+++ src/main/java/de/hbrs/model/Temperature.java
@@ -9,7 +9,7 @@ import org.springframework.data.relational.core.mapping.Table;
 @Table("temperatures")
 public record Temperature(
     @Id UUID id,
-    float temperature, // TODO: why float? Does the API only provide float precision? If not -> double
+    double temperature,
     String location,
     LocalDateTime time
 ) { }
